+++ zywienie-dla-zdrowia.php
@@ -54,11 +54,14 @@ require_once __DIR__ . '/includes/class-zfdz-wordpress-lab-result-catalog-servic
 require_once __DIR__ . '/includes/class-zfdz-wordpress-lab-result-public-presentation-result.php';
 require_once __DIR__ . '/includes/class-zfdz-wordpress-lab-result-public-presentation-resolver.php';
 require_once __DIR__ . '/includes/class-zfdz-wordpress-lab-result-public-presentation-service.php';
+require_once __DIR__ . '/includes/class-zfdz-wordpress-lab-result-public-url-resolver.php';
+require_once __DIR__ . '/includes/class-zfdz-wordpress-lab-result-shortcode.php';
 require_once __DIR__ . '/includes/class-zfdz-wordpress-menu-shortcode.php';
 require_once __DIR__ . '/includes/class-zfdz-wordpress-admin-status-page.php';
 require_once __DIR__ . '/includes/class-zfdz-wordpress-admin.php';
 
 ZFDZ_WordPress_Menu_Shortcode::register();
+ZFDZ_WordPress_Lab_Result_Shortcode::register();
 ZFDZ_WordPress_Admin::register();
 
 register_activation_hook(